clean: remove cli/readline.go
