+++ cli/commands/utils.go
@@ -2,6 +2,7 @@ package commands
 
 import (
 	"fmt"
+	"os"
 	"strings"
 )
 
@@ -30,4 +31,20 @@ var HelperSubCommand = &Definition{
 
 		fmt.Printf("%s\n", args[0].Def.StringRecurse(filter...))
 	},
+}
+
+func FileComplete(filePath *string) func(string) []string {
+	return func(line string) []string {
+		// var parts = strings.Split(line, " ")
+		var names = make([]string, 0)
+		var files, _ = os.ReadDir(*filePath)
+
+		// fmt.Printf("\n%s/%s\n", *filePath, parts[len(parts)-1])
+
+		for _, file := range files {
+			names = append(names, file.Name())
+		}
+
+		return names
+	}
 }
